+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,19 @@
 -->
 
 ## Upcoming Release
+* Added SuppressDiagnostics Parameter to Test-Deployment cmdlets.
+* Added PotentialChanges to WhatIf result.
+
+## Version 7.10.0
+* Fixed the issue that Get-AzReource not working with `-ExpandProperties`. [#11248]
+* Updated Resources SDK to 2024-11-01.
+* Added breaking change announcement for the following cmdlets due to API version for resource type may change.
+    - `Get-AzResource`
+    - `New-AzResource`
+    - `Set-AzResource`
+    - `Remove-AzResource`
+    - `Invoke-AzResourceAction`
+* Added ValidationLevel Parameter to WhatIf and Validate cmdlets for deployments.
 
 ## Version 7.9.0
 * Added `-ApplicationId` as an alias of `-ServicePrincipalName`.